Implement user creation (#34)
* Add docker build for UI.

Work on #15

* Add docker file for BFF.

Work on #15

* Add docker compose to configure entire stack.

Work on #15

* Add application configuration for UI.

Work on #15

* Use configured BFF address for calling API endpoints.

Work on #15.

* Provide configuration on local environment.

Work on #15

* Add scripts for local build of docker images.

* Add build script to solution file.

Work on #15

* Add keycloak containers.

Work on #15

* Add user creation page.

Work on #15

* Add placeholder for user creation in API.

Work on #15.

* Start adding user creation on API side.

Add aggregate root.
Add identity provider client abstraction.
Start preparing to connect to KeyCloak from API.

Work on #15

* Start preparing communication with Keycloak.

Configure HTTP client.
Try retrieving service account token.

Work on #15.

* Add user creation initial implementation.

Add storage of user and call to Keycloak to create credentials.

Work on #15

* Make e-mail address a value object.

Work on #15

* Clean-up naming convention.

In domain single namespace for single aggregate.
Aggregate has its own events (ES).

Work on #15

* Post review fixes.

Work on #15